Move the prospector into input
