+++ filebeat/include/list.go
@@ -1,16 +1,16 @@
 /*
-Package include imports all prospector packages so that they register
+Package include imports all input packages so that they register
 their factories with the global registry. This package can be imported in the
-main package to automatically register all of the standard supported prospectors
+main package to automatically register all of the standard supported inputs
 modules.
 */
 package include
 
 import (
 	// This list is automatically generated by `make imports`
-	_ "github.com/elastic/beats/filebeat/prospector/docker"
-	_ "github.com/elastic/beats/filebeat/prospector/log"
-	_ "github.com/elastic/beats/filebeat/prospector/redis"
-	_ "github.com/elastic/beats/filebeat/prospector/stdin"
-	_ "github.com/elastic/beats/filebeat/prospector/udp"
+	_ "github.com/elastic/beats/filebeat/input/docker"
+	_ "github.com/elastic/beats/filebeat/input/log"
+	_ "github.com/elastic/beats/filebeat/input/redis"
+	_ "github.com/elastic/beats/filebeat/input/stdin"
+	_ "github.com/elastic/beats/filebeat/input/udp"
 )
