+++ filebeat/input/log/prospector_windows_test.go
