+++ filebeat/input/log/stdin.go
