+++ filebeat/input/redis/_meta/Dockerfile
