+++ filebeat/input/redis/_meta/env
