+++ filebeat/input/redis/config.go
