+++ filebeat/input/redis/doc.go
