+++ filebeat/input/redis/harvester.go
