+++ filebeat/input/redis/prospector.go
